backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.3
python-multipart==0.0.9
pypdf2==3.0.1
pytesseract==0.3.10
anthropic==0.18.1
sqlalchemy==2.0.27
pytest==7.4.3
httpx==0.27.0
python-dotenv==1.0.1
pillow==10.2.0 